+++ packages/eslint-plugin-naming-convention/src/utils/is-jsx-file.ts
@@ -1,3 +1,0 @@
-import { JSX_EXTENSIONS } from "@eslint-react/shared";
-
-export const isJSXFile = (ext: string): ext is ".jsx" | ".tsx" => JSX_EXTENSIONS.includes(ext);