+++ packages/shared/src/constants.ts
@@ -4,8 +4,6 @@ export const GITHUB_URL = "https://github.com/rel1cx/eslint-react/blob/main";
 
 export const WEBSITE_URL = "https://eslint-react.rel1cx.io";
 
-export const JSX_EXTENSIONS = [".jsx", ".tsx"];
-
 // @see https://github.com/facebook/react/blob/6db7f4209e6f32ebde298a0b7451710dd6aa3e19/packages/react-dom-bindings/src/shared/sanitizeURL.js#L22
 // dprint-ignore
 // eslint-disable-next-line no-control-regex